# Conflicts:
#	nimare/meta/cbma/kernel.py
#	nimare/version.py

+++ nimare/utils.py
@@ -169,7 +169,7 @@ def mm2vox(xyz, affine):
     From here:
     http://blog.chrisgorgolewski.org/2014/12/how-to-convert-between-voxel-and-mm.html
     """
-    ijk = nib.affines.apply_affine(np.linalg.inv(affine), xyz)
+    ijk = nib.affines.apply_affine(np.linalg.inv(affine), xyz).astype(int)
     return ijk
